Move Console related stuff in the Console namespace
