+++ fixtures/Console/Command/TestConfigurable.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace KevinGH\Box\Command;
+namespace KevinGH\Box\Console\Command;
 
 class TestConfigurable extends Configurable
 {
