+++ src/Console/Command/ChangeableWorkingDirectory.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace KevinGH\Box\Command;
+namespace KevinGH\Box\Console\Command;
 
 use Assert\Assertion;
 use Symfony\Component\Console\Exception\RuntimeException;
