+++ src/Console/Logger/BuildLogger.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace KevinGH\Box\Logger;
+namespace KevinGH\Box\Console\Logger;
 
 use InvalidArgumentException;
 use Symfony\Component\Console\Output\OutputInterface;
