+++ src/Console/Logger/UpdateConsoleLogger.php
@@ -12,7 +12,7 @@ declare(strict_types=1);
  * with this source code in the file LICENSE.
  */
 
-namespace KevinGH\Box\Logger;
+namespace KevinGH\Box\Console\Logger;
 
 use Symfony\Component\Console\Style\SymfonyStyle;
 use Throwable;
