+++ tests/ConfigurationTest.php
@@ -19,8 +19,9 @@ use Generator;
 use Herrera\Annotations\Tokenizer;
 use InvalidArgumentException;
 use KevinGH\Box\Compactor\DummyCompactor;
+use KevinGH\Box\Compactor\InvalidCompactor;
 use KevinGH\Box\Compactor\Php;
-use KevinGH\Box\Helper\ConfigurationHelper;
+use KevinGH\Box\Console\ConfigurationHelper;
 use KevinGH\Box\Json\JsonValidationException;
 use Phar;
 use PHPUnit\Framework\TestCase;
